chore: move navigation folder
